Packing List Finish

+++ app/Http/Controllers/Api/StyleApiController.php
@@ -63,44 +63,49 @@ class StyleApiController extends Controller
             $size_size = strtoupper(request()['size_' . $times[$y]]);
             $weight = request()['weight_' . $times[$y]];
 
-            for($x = 0; $x <= 2; $x++){
+            if(($size_size !== null && $size_size !== "") && ($weight !== null && $weight !== "") ){
 
-                $carton_id = request()['carton_id_' . $places[$x] . '_' . $times[$y]];
-                $mcq = request()['mcq_'  . $places[$x] . '_' . $times[$y]];
+                for($x = 0; $x <= 2; $x++){
 
-                if($mcq !== null && $mcq !== ""){
+                    $carton_id = request()['carton_id_' . $places[$x] . '_' . $times[$y]];
+                    $mcq = request()['mcq_'  . $places[$x] . '_' . $times[$y]];
 
-                    if (Size::where('size', $size_size)->doesntExist()) {
-                        $size = Size::create(['size' => $size_size]);
-                    } else {
-                        $size = Size::where('size', $size_size)->first();
+                    if(($mcq !== null && $mcq !== "") && ($carton_id !== null && $carton_id !== "")){
+
+                        if (Size::where('size', $size_size)->doesntExist()) {
+                            $size = Size::create(['size' => $size_size]);
+                        } else {
+                            $size = Size::where('size', $size_size)->first();
+                        }
+
+                        $size_style_carton_id = $size->id . $style->id . $carton_id;
+
+                        if ($size->styles()->where('size_style_carton_id', $size_style_carton_id)->exists()) {
+
+                            $size->styles()->wherePivot('size_style_carton_id', '=', $size_style_carton_id)->detach();
+
+                            $size->styles()->attach([$style->id => [
+                                'weight' => $weight,
+                                'carton_id' => $carton_id,
+                                'mcq' => $mcq,
+                                'size_style_carton_id' => $size_style_carton_id,
+                            ]]);
+
+                        }else{
+                            $size->styles()->attach([$style->id => [
+                                'weight' => $weight,
+                                'carton_id' => $carton_id,
+                                'mcq' => $mcq,
+                                'size_style_carton_id' => $size_style_carton_id,
+                            ]]);
+                        }
+
+
                     }
 
-                    $size_style_carton_id = $size->id . $style->id . $carton_id;
 
-                    if ($size->styles()->where('size_style_carton_id', $size_style_carton_id)->exists()) {
-
-                        $size->styles()->wherePivot('size_style_carton_id', '=', $size_style_carton_id)->detach();
-
-                        $size->styles()->attach([$style->id => [
-                            'weight' => $weight,
-                            'carton_id' => $carton_id,
-                            'mcq' => $mcq,
-                            'size_style_carton_id' => $size_style_carton_id,
-                        ]]);
-
-                    }
-                    $size->styles()->attach([$style->id => [
-                        'weight' => $weight,
-                        'carton_id' => $carton_id,
-                        'mcq' => $mcq,
-                        'size_style_carton_id' => $size_style_carton_id,
-                    ]]);
 
                 }
-
-
-
             }
